Extracting directory mover function to file
To better assist in finding the module, moving it to its "conventional"
location.
